refactor(fe, dashboard): 비어있는 테이블 데이터 처리
